Removes old sava data and replaces with the new one
