+++ Danki2/Assets/Prefabs/Meta/MapGenerator.prefab
@@ -9,7 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 3011644956931086658}
-  - component: {fileID: 3011644956931086685}
+  - component: {fileID: 1984440938464060650}
   m_Layer: 0
   m_Name: MapGenerator
   m_TagString: Untagged
@@ -31,7 +31,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &3011644956931086685
+--- !u!114 &1984440938464060650
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -40,6 +40,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 3011644956931086684}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 85c469d62d60490b886b642eb2f0b015, type: 3}
+  m_Script: {fileID: 11500000, guid: 6d13d8e300f2fa44093ac92b8c3e516e, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
