+++ Danki2/Assets/Prefabs/Meta/NewSaveGenerator.prefab
@@ -9,7 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 5168182131318211707}
-  - component: {fileID: 5168182131318211704}
+  - component: {fileID: 3907087133687394453}
   m_Layer: 0
   m_Name: NewSaveGenerator
   m_TagString: Untagged
@@ -31,7 +31,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &5168182131318211704
+--- !u!114 &3907087133687394453
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -40,6 +40,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 5168182131318211705}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2667ea8ebf2c42c49ef56c4cab431ca2, type: 3}
+  m_Script: {fileID: 11500000, guid: 79ec5556f0ae9f84a8599cb98ad8e07b, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
