+++ Danki2/Assets/Scripts/Actor/Spawner/PlayerSpawner.cs
@@ -10,7 +10,7 @@ public class PlayerSpawner : MonoBehaviour
 
     public void Spawn()
     {
-        Pole entranceSide = SceneLookup.Instance.GetEntranceSide(PersistenceManager.Instance.SaveData.CurrentRoomSaveData.Scene, id);
+        Pole entranceSide = SceneLookup.Instance.GetEntranceSide(PersistenceManager.Instance.SaveData.CurrentRoomNode.Scene, id);
         float yRotation = OrientationUtils.GetYRotation(OrientationUtils.GetReversedPole(entranceSide));
 
         spawner.Spawn(ActorType.Player).transform.Rotate(0, yRotation, 0);
