+++ Danki2/Assets/Scripts/Persistence/NewSaveGeneration/MapGenerator.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: 85c469d62d60490b886b642eb2f0b015
-timeCreated: 1615972856
+fileFormatVersion: 2
+guid: 6d13d8e300f2fa44093ac92b8c3e516e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
