--- Danki2/Assets/Scripts/Persistence/NewSaveGeneration/MapGenerator2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6d13d8e300f2fa44093ac92b8c3e516e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
