+++ Danki2/Assets/Scripts/Persistence/NewSaveGeneration/NewSaveGenerator.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: 2667ea8ebf2c42c49ef56c4cab431ca2
-timeCreated: 1615201092
+fileFormatVersion: 2
+guid: 79ec5556f0ae9f84a8599cb98ad8e07b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
