--- Danki2/Assets/Scripts/Persistence/NewSaveGeneration/NewSaveGenerator2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 79ec5556f0ae9f84a8599cb98ad8e07b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
