+++ Danki2/Assets/Scripts/Persistence/PersistenceManager.cs
@@ -1,4 +1,6 @@
-﻿/// <summary>
+﻿using UnityEngine;
+
+/// <summary>
 /// Handles persistence between gameplay rooms. The public property SaveData is available for anything to read from
 /// at any point so that it can initialise itself.
 ///
@@ -39,13 +41,17 @@ public class PersistenceManager : Singleton<PersistenceManager>
     /// The save data is updated before we transition, then the reference to the current room is updated so that
     /// we can save as if we were right at the start of the next room.
     /// </summary>
-    public virtual void TransitionToNextRoom(int nextRoomId)
+    public virtual void TransitionToNextRoom(RoomNode nextRoomNode)
     {
         UpdateSaveData();
-        SaveData.CurrentRoomId = nextRoomId;
-        // TODO: Call through to generate new layer... (or maybe just do that logic all here?)
+        SaveData.CurrentRoomNode = nextRoomNode;
+        
+        Random.state = SaveData.RandomState;
+        MapGenerator.Instance.GenerateNextLayer(SaveData.CurrentRoomNode);
+        SaveData.RandomState = Random.state;
+        
         SaveDataManager.Instance.Save(SaveData);
-        SceneUtils.LoadScene(SaveData.CurrentRoomSaveData.Scene);
+        SceneUtils.LoadScene(SaveData.CurrentRoomNode.Scene);
     }
 
     /// <summary>
@@ -54,9 +60,9 @@ public class PersistenceManager : Singleton<PersistenceManager>
     public virtual void TransitionToDefeatRoom()
     {
         UpdateSaveData();
-        SaveData.CurrentRoomId = SaveData.DefeatRoomId;
+        SaveData.CurrentRoomNode = SaveData.DefeatRoom;
         SaveDataManager.Instance.Save(SaveData);
-        SceneUtils.LoadScene(SaveData.DefeatRoomSaveData.Scene);
+        SceneUtils.LoadScene(SaveData.DefeatRoom.Scene);
     }
 
     private void UpdateSaveData()
@@ -65,24 +71,24 @@ public class PersistenceManager : Singleton<PersistenceManager>
         SaveData.SerializableAbilityTree = ActorCache.Instance.Player.AbilityTree.Serialize();
         SaveData.RuneSockets = ActorCache.Instance.Player.RuneManager.RuneSockets;
 
-        RoomSaveData currentRoomSaveData = SaveData.CurrentRoomSaveData;
-        switch (currentRoomSaveData.RoomType)
+        RoomNode currentRoomNode = SaveData.CurrentRoomNode;
+        switch (currentRoomNode.RoomType)
         {
             case RoomType.Combat:
             case RoomType.Boss:
-                CombatRoomSaveData combatRoomSaveData = currentRoomSaveData.CombatRoomSaveData;
+                CombatRoomSaveData combatRoomSaveData = currentRoomNode.CombatRoomSaveData;
                 combatRoomSaveData.EnemiesCleared = CombatRoomManager.Instance.EnemiesCleared;
                 break;
             case RoomType.Ability:
-                AbilityRoomSaveData abilityRoomSaveData = currentRoomSaveData.AbilityRoomSaveData;
+                AbilityRoomSaveData abilityRoomSaveData = currentRoomNode.AbilityRoomSaveData;
                 abilityRoomSaveData.AbilitiesViewed = AbilitySelectionRoomManager.Instance.AbilitiesViewed;
                 abilityRoomSaveData.AbilitySelected = AbilitySelectionRoomManager.Instance.AbilitySelected;
                 break;
             case RoomType.Healing:
-                currentRoomSaveData.HealingRoomSaveData.HasHealed = HealingRoomManager.Instance.HasHealed;
+                currentRoomNode.HealingRoomSaveData.HasHealed = HealingRoomManager.Instance.HasHealed;
                 break;
             case RoomType.Rune:
-                RuneRoomSaveData runeRoomSaveData = currentRoomSaveData.RuneRoomSaveData;
+                RuneRoomSaveData runeRoomSaveData = currentRoomNode.RuneRoomSaveData;
                 runeRoomSaveData.RunesViewed = RuneRoomManager.Instance.RunesViewed;
                 runeRoomSaveData.RuneSelected = RuneRoomManager.Instance.RuneSelected;
                 break;
