+++ Danki2/Assets/Scripts/Persistence/SaveData/SaveData.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: 5930728f55cb46a893e947f8948352d5
-timeCreated: 1614763848
+fileFormatVersion: 2
+guid: 8fd41bfcd3b0a314cb093c050e84a46f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
