+++ Danki2/Assets/Scripts/Persistence/SaveData/Serialization/SerializableRoomNode.cs
