--- Danki2/Assets/Scripts/Persistence/SaveData/Serialization/SerializableRoomSaveData.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 548cc8a1898f45c5a81017e701f17d39
-timeCreated: 1615197363
