--- Danki2/Assets/Scripts/Persistence/SaveData2.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c6349c50e7ffe614da4b3648a454b018
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
