--- Danki2/Assets/Scripts/Persistence/SaveData2/SaveData2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8fd41bfcd3b0a314cb093c050e84a46f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
