--- Danki2/Assets/Scripts/Persistence/SaveData2/Serialization.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8ad5891259204a9d931c8e8082acdfb2
-timeCreated: 1620815800
