+++ Danki2/Assets/Scripts/RoomManagement/Transitions/RoomTransitioner.cs
@@ -11,7 +11,7 @@ public class RoomTransitioner : MonoBehaviour
     
     private void Start()
     {
-        active = PersistenceManager.Instance.SaveData.CurrentRoomSaveData.RoomTransitionerIdToTransitionData.ContainsKey(id);
+        active = PersistenceManager.Instance.SaveData.CurrentRoomNode.ExitIdToChildLookup.ContainsKey(id);
         if (!active) return;
 
         GameplayRoomTransitionManager.Instance.CanTransitionSubject.Subscribe(HandleCanTransitionSubject);
@@ -23,8 +23,9 @@ public class RoomTransitioner : MonoBehaviour
         if (!GameplayRoomTransitionManager.Instance.CanTransition) return;
         if (transform.DistanceFromPlayer() > TransitionDistance) return;
 
-        int nextRoomId = PersistenceManager.Instance.SaveData.CurrentRoomSaveData.RoomTransitionerIdToTransitionData[id].NextRoomId;
-        PersistenceManager.Instance.TransitionToNextRoom(nextRoomId);
+        PersistenceManager.Instance.TransitionToNextRoom(
+            PersistenceManager.Instance.SaveData.CurrentRoomNode.ExitIdToChildLookup[id]
+        );
     }
 
     private void HandleCanTransitionSubject(bool canTransition)
