+++ Danki2/Assets/Scripts/Scenes/SceneManagers/GameplayEntrySceneManager.cs
@@ -4,6 +4,6 @@ public class GameplayEntrySceneManager : MonoBehaviour
 {
     private void Start()
     {
-        SceneUtils.LoadScene(PersistenceManager.Instance.SaveData.CurrentRoomSaveData.Scene);
+        SceneUtils.LoadScene(PersistenceManager.Instance.SaveData.CurrentRoomNode.Scene);
     }
 }
